Potential issue elimination
If cache refreshes whilst there are pending deletions during which a torrent is ran through the checker script, it will resee the same torrents that have been deleted. Better off deleting the torrents immediately after they are sent to the deleter script. And then queue the actual deletion of files.